Break spatial simulations
This is the first step towards cleaning up the
lattice types. Main points:

The `Site` type has been greatly simplified.
A `Site` does not carry references to neighbors.
Added a plotting recipe to treat these objects
as points in 1D, 2D, and 3D space.

The `Lattice` type replaces `SLattice` and
`NLattice`.
These objects will be the data stored during
simulation.
Added a plotting recipe to make visualization
as painless as possible.
Later we will add a way to label cell/particle
types automatically.

The new lean `Lattice` type removes a lot
of fields, notably all the dictionaries used
for search operations.
The goal is to use `searchsorted` and its
interface to do these operations because it
implements a binary search on arbitrary types
supporting `isless`.

Simplified all the neighborhood cruft.
Each `Lattice` object has a type parameter that
indicates its topology, the `VonNeumann` and
`Hexagonal` types.
Neighborhoods are dumb arrays.
The goal is to use `VonNeumann` and `Hexagonal`
to mediate changes to neighborhoods.

Finally, all the sample class logic is isolated
in a new `SampleClassEnumeration` type.
Along with `Lattice`, this new type will be
important in re-implementing all the look-ups
in updating sample classes and neighborhoods.