Add a workaround for broken integrations.
Calling `DataFrame(result)` does not build the correct `DataFrame`.
There is a documented issue in IterableTables.jl that suggests this is
due to a design choice in DataFrames.jl. The `tablefy` function is
our last resort if the correct table iterator fails.

Note: I am not sure whether `SamplePath` or `SamplePathIterator` is fed
through the IterableTables.jl pipeline. We should probably wrap results
in another type to remove this overlap, but it might not be a big deal.